+++ contracts/otcMaker/OtcMaker.sol
@@ -106,8 +106,8 @@ contract OtcMaker is SafeOwnable, EIP712Upgradeable, IOtcMaker, OtcMakerStorageV
         JitLiquidityParams calldata jitLiquidityParams,
         bytes calldata signature
     ) external override onlyCaller returns (int256, int256) {
-        // OM_NLO: not limit order
-        require(limitOrderParams.orderType == ILimitOrderBook.OrderType.LimitOrder, "OM_NLO");
+        // OM_NOMO: not otc maker order
+        require(limitOrderParams.orderType == ILimitOrderBook.OrderType.OtcMakerOrder, "OM_NOMO");
 
         // IAccountBalance -> get before takerPositionSize, takerPositionNotional
         AccountMarket.Info memory accountInfoBefore = IAccountBalance(_accountBalance).getAccountInfo(