+++ test/foundry/OtcMaker/OtcMaker.openPositionFor.t.sol
@@ -30,7 +30,7 @@ contract OtcMakerOpenPositionForTest is OtcMakerSetup, EIP712Upgradeable {
         );
 
         ILimitOrderBook.LimitOrder memory limitOrderParams = _generateLimitOrderParams(
-            ILimitOrderBook.OrderType.LimitOrder,
+            ILimitOrderBook.OrderType.OtcMakerOrder,
             alice,
             address(perp.baseToken()),
             false,
@@ -88,7 +88,7 @@ contract OtcMakerOpenPositionForTest is OtcMakerSetup, EIP712Upgradeable {
         );
 
         ILimitOrderBook.LimitOrder memory limitOrderParams = _generateLimitOrderParams(
-            ILimitOrderBook.OrderType.LimitOrder,
+            ILimitOrderBook.OrderType.OtcMakerOrder,
             alice,
             address(perp.baseToken()),
             false,
@@ -168,7 +168,7 @@ contract OtcMakerOpenPositionForTest is OtcMakerSetup, EIP712Upgradeable {
 
         bytes memory signature = _signLimitOrderParams(alicePrivateKey, limitOrderParams);
 
-        vm.expectRevert(bytes("OM_NLO"));
+        vm.expectRevert(bytes("OM_NOMO"));
         vm.prank(otcMakerCaller);
         otcMaker.openPositionFor(
             limitOrderParams,